+++ src/main/java/com/Mybatis/Mybatis/module/repository/impl/RepositoryUsuarioimpl.java
@@ -3,12 +3,8 @@ package com.Mybatis.Mybatis.module.repository.impl;
 import com.Mybatis.Mybatis.module.dao.UsuarioService;
 import com.Mybatis.Mybatis.module.entities.Usuario;
 import com.Mybatis.Mybatis.module.repository.RepositoryUsuario;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-
-
 import javax.annotation.Resource;
-import java.util.Collections;
 import java.util.List;
 
 @Repository("repositoryUsuario")
@@ -23,7 +19,7 @@ public class RepositoryUsuarioimpl implements RepositoryUsuario {
     }
 
     @Override
-    public Usuario getUsuarioById(int id) {
+    public Usuario getUsuarioById(long id) {  // Cambiado a long
         return usuarioService.getUsuarioById(id);
     }
 
@@ -38,9 +34,10 @@ public class RepositoryUsuarioimpl implements RepositoryUsuario {
     }
 
     @Override
-    public void deleteUsuarioById(int id) {
+    public void deleteUsuarioById(long id) {  // Cambiado a long
         usuarioService.deleteUsuarioById(id);
     }
 }
+
